+++ CHANGELOG.md
@@ -10,13 +10,25 @@ any parts of the framework not mentioned in the documentation should generally b
 
 ## [Unreleased]
 
-### Changed
+### Added
 
-* Allowed repeated filter query parameters.
+* Added support for serializing nested serializers as attribute json value introducing setting `JSON_API_SERIALIZE_NESTED_SERIALIZERS_AS_ATTRIBUTE`
 
 ### Fixed
 
-* Avoid `AttributeError` for PUT and PATCH methods when using `APIView` 
+* Avoid `AttributeError` for PUT and PATCH methods when using `APIView`
+
+### Changed
+
+* `SerializerMethodResourceRelatedField` is now consistent with DRF `SerializerMethodField`:
+   * Pass `method_name` argument to specify method name. If no value is provided, it defaults to `get_{field_name}`
+* Allowed repeated filter query parameters.
+
+### Deprecated
+
+* Deprecate `source` argument of `SerializerMethodResourceRelatedField`, use `method_name` instead
+* Rendering nested serializers as relationships is deprecated. Use `ResourceRelatedField` instead
+
 
 ## [3.1.0] - 2020-02-08